pyproject.toml
[project]
name = "oc6-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "imbalanced-learn>=0.14.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.8.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "xgboost>=3.1.2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic-settings>=2.7.0",
    "httpx>=0.28.0",
    "joblib>=1.4.0",
    "streamlit>=1.41.0",
    "plotly>=5.10.0,<6",
    "evidently>=0.7.0",
    "psycopg2-binary>=2.9.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.0",
]
